+++ CHANGELOG.md
@@ -14,6 +14,8 @@ The format mostly follows [Keep a Changelog](http://keepachangelog.com/en/1.0.0/
 - Documentation now has a section on the configuration settings (`--edit-config`)
 - IFTTT reporter (#512, by Florian Gaultier)
 - New filter: ``ocr`` to convert text in images to plaintext (using Tesseract OCR)
+- Add `--test-diff-filter` command-line option that can test the `diff_tool` and
+  `diff_filter` settings for a job (needs at least 2 historic snapshots)
 
 ### Changed